# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Semantic engine
  version: "0.1.0"
  description: |
    An Apache Ossie semantic model compiled to governed, dialect-correct SQL.

    There is no endpoint that accepts SQL, and there will not be one. The only
    expressible request is a semantic one: metrics, dimensions and structured
    filters. That is what makes an ungoverned query inexpressible rather than
    merely discouraged.

    Column access is resolved before SQL is emitted. A denied request returns a
    refusal with a reason; it never returns an empty result set, because a
    silent empty answer trains a caller to report a wrong number confidently.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://127.0.0.1:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /v1/compile:
    post:
      tags: [query]
      operationId: compile
      summary: Return the SQL a request compiles to, without running it
      description: |
        A dry run. The governance gate runs exactly as it does for a query, so
        this is not a way to read the SQL for a metric you may not query: a
        denied request returns a refusal and no SQL.

        The decision is audited as `compiled` rather than `allowed`, which keeps
        inspecting the model distinguishable from executing against it.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QueryRequest" }
      responses:
        "200":
          description: Compiled SQL
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CompileResult" }
        "400": { $ref: "#/components/responses/Refused" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Refused" }
        "404": { $ref: "#/components/responses/Refused" }
        "422": { $ref: "#/components/responses/Refused" }
        "503": { $ref: "#/components/responses/Refused" }

  # Asynchronous execution. A warehouse query can run longer than the default
  # HTTP timeout of every agent framework in use; synchronously the client
  # gives up, retries, and the warehouse runs and bills the query twice.
  /v1/jobs:
    post:
      tags: [query]
      operationId: submitJob
      summary: Start a query in the background and return a job identifier
      description: |
        Compilation, and therefore the governance gate, happens synchronously
        before the job exists. A request the caller may not make never becomes
        a job: it is refused on this call. A 202 is therefore a promise that
        the query was authorized, not merely that it was accepted.

        The compiled SQL is returned immediately, so a caller can show its work
        without waiting for the warehouse.

        A job is visible only to the identity that submitted it.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QueryRequest" }
      responses:
        "202":
          description: Accepted and running
          headers:
            Location:
              description: The job resource to poll.
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobAccepted" }
        "400": { $ref: "#/components/responses/Refused" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Refused" }
        "404": { $ref: "#/components/responses/Refused" }
        "422": { $ref: "#/components/responses/Refused" }
        "503": { $ref: "#/components/responses/Refused" }
  /v1/jobs/{id}:
    get:
      tags: [query]
      operationId: getJob
      summary: Poll a job and collect a page of its rows
      description: |
        Returns 200 whenever the job is readable by this caller, including when
        the job itself failed: the poll succeeded, and the failure is described
        in the body with the same `code`, `hint` and `retry` a synchronous
        refusal carries. Branch on `state`, not on the status code.

        `row_count` is the size of the whole result rather than of the page, so
        a caller knows how much remains before walking it. When `next_cursor`
        is present there are more rows; pass it back to read the next page.

        A job belonging to another identity returns 404 rather than 403, so
        job identifiers cannot be probed for existence.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: cursor
          in: query
          required: false
          description: Opaque cursor from a previous page's `next_cursor`.
          schema: { type: string }
        - name: page_size
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 10000, default: 1000 }
      responses:
        "200":
          description: Job state, and rows once it has succeeded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobStatus" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/Refused" }
        "422": { $ref: "#/components/responses/Refused" }
    delete:
      tags: [query]
      operationId: cancelJob
      summary: Cancel a running query
      description: |
        Idempotent. Cancelling a job that already finished returns its existing
        state rather than an error.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: The job state after cancellation
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id: { type: string }
                  state: { $ref: "#/components/schemas/JobState" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/Refused" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        The token identifies a workload, and that workload identity is what the
        governance gate and the executor use. Grants go to service accounts and
        groups, never to individual humans.
  responses:
    Unauthorized:
      description: Missing or unrecognised credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Refusal" }
    Refused:
      description: |
        The engine understood the request and declined it. Branch on `retry`.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Refusal" }
  schemas:
...
    Refusal:
      type: object
      required: [code, reason, retry]
      properties:
        code:
          type: string
          description: Machine-readable reason. Branch on this, not on the text.
          examples: [fan_out_would_inflate, access_denied, unknown_metric]
        reason:
          type: string
          description: One sentence stating what was wrong.
        hint:
          type: string
          description: |
            What to do instead. For a fan-out refusal this names the metrics
            defined at the grain where the question is well defined.
        retry: { $ref: "#/components/schemas/Retry" }
    Retry:
      type: string
      enum: [modify, later, never]
      description: |
        What a caller should do about a refusal. This is the field an agent
        branches on; without it a caller either gives up on a typo or loops
        forever on a denial.

        `modify`: the request is answerable but not as written. Change the
        arguments and try again; repeating it unchanged will not work.

        `later`: something outside the request failed, such as the policy
        source being unreachable, or the caller is at their concurrency limit.
        The same request may succeed later.

        `never`: no version of this request from this caller will succeed.
        Stop, and say so rather than guessing.
    Filter:
      type: object
      required: [dimension, op]
      description: |
        A structured predicate. Never a SQL fragment: that is what keeps an
        ungoverned predicate inexpressible.
      properties:
        dimension:
          type: string
          description: "`dataset.field`, or `namespace.dataset.field`."
        op:
          type: string
          enum: [eq, ne, in, not_in, gt, gte, lt, lte, between, is_null, is_not_null]
        values:
          type: array
          description: |
            Value count depends on the operator: one for the comparisons, two
            for `between`, one or more for `in` and `not_in`, none for the null
            checks. Values are validated against the dimension's declared type
            before planning, and are always bound as parameters.
          items:
            anyOf:
              - { type: string }
              - { type: number }
              - { type: boolean }
    Order:
      type: object
      required: [field]
      properties:
        field:
          type: string
          description: "An output column name: a metric name or a dimension field name."
        desc: { type: boolean, default: false }
    QueryRequest:
      type: object
      required: [metrics]
      properties:
        metrics:
          type: array
          minItems: 1
          items: { type: string }
          description: |
            Qualified as `namespace.metric`, or bare when unambiguous. Metrics
            at different grains, or in different namespaces, are aggregated
            separately and joined on the shared dimensions.
        dimensions:
          type: array
          items: { type: string }
          description: "`dataset.field`, or `namespace.dataset.field`."
        filters:
          type: array
          items: { $ref: "#/components/schemas/Filter" }
        grain:
          type: string
          enum: [second, minute, hour, day, week, month, quarter, year]
          description: |
            Time bucket applied to the selected time dimension. Requesting a
            grain with no time dimension selected is refused rather than ignored.
        limit:
          type: integer
          minimum: 1
          maximum: 100000
          default: 1000
        order_by:
          type: array
          items: { $ref: "#/components/schemas/Order" }
...
    CompileResult:
      type: object
      required: [compiled_sql, columns, dialect, model_version]
      properties:
        compiled_sql: { type: string }
        columns: { type: array, items: { type: string } }
        dialect: { type: string }
        namespace: { type: string }
        parts:
          type: integer
          description: |
            How many facts were aggregated separately and joined. More than one
            means the query spanned grains or namespaces.
        model_version: { type: string }
...
    JobState:
      type: string
      enum: [running, succeeded, failed, cancelled]
      description: |
        `running` is the only non-terminal state. `cancelled` is distinct from
        `failed`: the caller stopped it, the warehouse did not fail.
    JobAccepted:
      type: object
      description: |
        The acknowledgement of a submitted job. It carries the compiled SQL
        because compilation already happened, synchronously, before the 202.
      required: [job_id, state, compiled_sql]
      properties:
        job_id: { type: string, description: Opaque identifier, 128 bits of randomness. }
        state: { $ref: "#/components/schemas/JobState" }
        submitted_at: { type: string, format: date-time }
        compiled_sql: { type: string }
        columns:
          type: array
          items: { type: string }
        namespace: { type: string }
        model_version: { type: string }
        dialect: { type: string }
        parts:
          type: integer
          description: How many facts were aggregated separately and joined.
    JobStatus:
      type: object
      description: |
        The result of polling a job. Which fields are present depends on
        `state`: rows only when `succeeded`, refusal fields only when `failed`.
      required: [job_id, state]
      properties:
        job_id: { type: string }
        state: { $ref: "#/components/schemas/JobState" }
        submitted_at: { type: string, format: date-time }
        finished_at:
          type: string
          format: date-time
          description: Present once the job reached a terminal state.
        elapsed_ms:
          type: integer
          description: Present while running.
        duration_ms:
          type: integer
          description: Present once finished.
        columns:
          type: array
          items: { type: string }
        rows:
          type: array
          description: One page of rows. Present only when `state` is `succeeded`.
          items:
            type: array
            items: {}
        row_count:
          type: integer
          description: Rows in the whole result, not in this page.
        next_cursor:
          type: string
          description: Present when more rows remain. Pass it back as `cursor`.
        compiled_sql: { type: string }
        namespace: { type: string }
        model_version: { type: string }
        dialect: { type: string }
        code:
          type: string
          description: Present only when `state` is `failed`.
        reason: { type: string }
        hint: { type: string }
        retry: { $ref: "#/components/schemas/Retry" }